Refactored code generation to make it easier to build upon in future.
1. Renamed CodeGenerator to Emitter and moved to Emit folder.
2. Split Emitter into multiple files for easier maintenance.
3. Converted namespaces to file scope.
4. Removed unused namespace imports.
5. Renamed all serializer hosts to more meaningful terms.
