+++ FJS.Generator/Emit/Emitter.Scalar.cs
@@ -1,0 +1,72 @@
+using System.Collections.Generic;
+using FJS.Generator.Model;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
+using static Microsoft.CodeAnalysis.CSharp.SyntaxKind;
+
+namespace FJS.Generator.Emit;
+
+static partial class Emitter
+{
+    static void WritePrimitive(List<StatementSyntax> stmts, MemberData member)
+    {
+        stmts.Add(
+            ExpressionStatement(
+                InvocationExpression(
+                    MemberAccessExpression(SimpleMemberAccessExpression,
+                        IdentifierName("writer"),
+                        IdentifierName(member.MemberType == MemberType.String ? "WriteString" : "WriteNumber")),
+                    ArgumentList(SeparatedList(
+                        [
+                            Argument(LiteralExpression(StringLiteralExpression, Literal(member.Name))),
+                                        Argument(MemberAccessExpression(SimpleMemberAccessExpression,
+                                            IdentifierName("obj"),
+                                            IdentifierName(member.Name))),
+                        ])))));
+    }
+
+    static void WriteNullable(List<StatementSyntax> stmts, MemberData member)
+    {
+        stmts.Add(
+            IfStatement(
+                MemberAccessExpression(SimpleMemberAccessExpression,
+                    MemberAccessExpression(SimpleMemberAccessExpression,
+                        IdentifierName("obj"),
+                        IdentifierName(member.Name)),
+                    IdentifierName("HasValue")),
+                Block(new StatementSyntax[]
+                    {
+                        ExpressionStatement(
+                            InvocationExpression(
+                                MemberAccessExpression(SimpleMemberAccessExpression,
+                                    IdentifierName("writer"),
+                                    IdentifierName("WritePropertyName")),
+                                ArgumentList(SeparatedList(
+                                    [
+                                        Argument(LiteralExpression(StringLiteralExpression, Literal(member.Name)))
+                                    ])))),
+                        ExpressionStatement(
+                            InvocationExpression(
+                                MemberAccessExpression(SimpleMemberAccessExpression,
+                                    IdentifierName("writer"),
+                                    IdentifierName("WritePropertyName")),
+                                ArgumentList(SeparatedList(
+                                    [
+                                        Argument(LiteralExpression(StringLiteralExpression, Literal(member.Name)))
+                                    ]))))
+                    }),
+                ElseClause(
+                    Block(new StatementSyntax[]
+                    {
+                        ExpressionStatement(
+                            InvocationExpression(
+                                MemberAccessExpression(SimpleMemberAccessExpression,
+                                    IdentifierName("writer"),
+                                    IdentifierName("WriteNull")),
+                                ArgumentList(SeparatedList(
+                                    [
+                                        Argument(LiteralExpression(StringLiteralExpression, Literal(member.Name)))
+                                    ]))))
+                    }))));
+    }
+}
