+++ FJS.Generator/Model/CodeGeneratorState.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace FJS.Generator.Model;
+
+class CodeGeneratorState
+{
+    public List<TypeData> TypesToGenerate { get; } = new();
+    public HashSet<string> Processed { get; } = new();
+}
