+++ FJS.Generator/Model/Host.cs
@@ -1,13 +1,12 @@
 using System.Collections.Generic;
 
-namespace FJS.Generator.Model
+namespace FJS.Generator.Model;
+
+class Host
 {
-    class Host
-    {
-        public string Name { get; set; }
+    public string Name { get; set; }
 
-        public string Namespace { get; set; }
+    public string Namespace { get; set; }
 
-        public List<TypeData> Types { get; } = new();
-    }
+    public List<TypeData> Types { get; } = new();
 }
