+++ FJS.Generator/Model/MemberData.cs
@@ -1,15 +1,14 @@
-namespace FJS.Generator.Model
+namespace FJS.Generator.Model;
+
+class MemberData
 {
-    class MemberData
-    {
-        public string Name { get; set; }
+    public string Name { get; set; }
 
-        public bool CanRead { get; set; }
+    public bool CanRead { get; set; }
 
-        public bool CanWrite { get; set; }
+    public bool CanWrite { get; set; }
 
-        public MemberTypes MemberType { get; set; }
+    public MemberType MemberType { get; set; }
 
-        public TypeData CollectionElementType { get; set; }
-    }
+    public TypeData CollectionElementType { get; set; }
 }
