+++ FJS.Generator/Model/MemberType.cs
@@ -1,0 +1,12 @@
+namespace FJS.Generator.Model;
+
+enum MemberType
+{
+    Unspecified,
+    Number,
+    String,
+    ComplexObject,
+    SequentialCollection,
+    AssociativeCollection,
+    Nullable,
+}
