--- FJS.Generator/Model/MemberTypes.cs
@@ -1,13 +1,0 @@
-namespace FJS.Generator.Model
-{
-    enum MemberTypes
-    {
-        Unspecified,
-        Number,
-        String,
-        ComplexObject,
-        SequentialCollection,
-        AssociativeCollection,
-        Nullable,
-    }
-}
