+++ FJS.UnitTests/Primitives/NullableFields.cs
@@ -1,4 +1,3 @@
-using FJS.Common.Metadata;
 using System.Text;
 
 namespace FJS.UnitTests.Primitives;
